requirements.txt
openai
stockfish
numpy
safe-cast
chess
nltk
continuous-threading
sounddevice
soundfile
pytermgui
pyperclip
PySide6
darkdetect
win32mica
pyttsx3
pyserial